+++ Tests/Resources/app/config/config.php
@@ -1,7 +1,4 @@
 <?php
-// Tests/Resources/app/config/config.php
 
 $loader->import(CMF_TEST_CONFIG_DIR.'/default.php');
-$loader->import(CMF_TEST_CONFIG_DIR.'/phpcr_odm.php');
-$loader->import(CMF_TEST_CONFIG_DIR.'/sonata_admin.php');
 $loader->import(__DIR__.'/cmf_seo.yml');