+++ Tests/Resources/app/config/config_phpcr.php
@@ -1,0 +1,5 @@
+<?php
+
+$loader->import(CMF_TEST_CONFIG_DIR.'/phpcr_odm.php');
+$loader->import(CMF_TEST_CONFIG_DIR.'/sonata_admin.php');
+$loader->import(__DIR__.'/cmf_seo.phpcr.yml');