+++ Tests/WebTest/Admin/SeoContentAdminExtensionTest.php
@@ -66,12 +66,11 @@ class SeoContentAdminExtensionTest extends BaseTestCase
 
     public function testExtraPropertyEditView()
     {
-        $crawler = $this->client->request('GET', '/admin/cmf/seo/seoawarecontent/test/content/content-extra-property/edit');
+        $crawler = $this->client->request('GET', '/admin/cmf/seo/seoawarecontent/test/content/content-extra/edit');
         $res = $this->client->getResponse();
 
         $this->assertEquals(200, $res->getStatusCode());
         $this->assertCount(1, $crawler->filter('html:contains("Key")'));
         $this->assertCount(1, $crawler->filter('html:contains("Value")'));
-        $this->assertCount(1, $crawler->filter('html:contains("Type")'));
     }
 }